engine/requirements.txt
asyncio
pika
python-binance